Resolve numba error and compile _get_area in nopython mode
The _get_area function was getting compiled in object mode because it
was using the builtin `abs` function but calling it on the result of a
calculation operating on a numpy array, which numba didn't think made sense.
Changing it to `np.abs` allows it to compile correctly in nopython mode.
It might or might not be faster, but at least this gets rid of the rather
long warning message.

This also changes the decorator from 'jit' to `njit` to force nopython
mode, just for explicitness and because there's no reason these functions
shouldn't be nopython compatible, so if they stop compiling successfully
that's more likely to be a bug than a reason to fall back to object mode.